[FE] feat: 상품 정보 이미지 수정 스펙 추가 #102
