+++ frontend/src/components/Inputs/ImageInput.vue
@@ -23,7 +23,6 @@
 
 <script lang="ts">
 import { Component, Vue, Ref, Prop } from 'vue-property-decorator';
-import axios from 'axios';
 import Icon from '@components/Common/Icon.vue';
 
 function generateRandomString(length = 10, acc = '') {
@@ -36,30 +35,20 @@ function generateRandomString(length = 10, acc = '') {
   return generateRandomString(length, acc);
 }
 
-async function urlToFileObject(url: string) {
-  const { data } = await axios.get(url, { responseType: 'blob' });
-  return new File([data], 'image.png', { type: data.type });
-}
-
-interface IInitValue {
-  id: number;
-  path: string;
-}
-
 @Component({
   components: { Icon },
 })
 export default class ImageInput extends Vue {
   @Ref() fileInput!: HTMLInputElement;
 
-  @Prop() initValues!: IInitValue[];
+  @Prop() initValues!: File[];
 
   private imageURLMap = new Map<string, File>();
 
   public async created() {
     for (let i = 0, len = this.initValues.length; i < len; i += 1) {
       const hashcode = generateRandomString(10);
-      const file = await urlToFileObject(this.initValues[i].path);
+      const file = this.initValues[i];
 
       if (!this.imageURLMap.has(hashcode)) {
         this.imageURLMap.set(hashcode, file);
